Factor out createTrieDataStructure().
